% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.fpr.table.R
\name{extract.fpr.table}
\alias{extract.fpr.table}
\title{Extract dataframes that contain the pairwise FPR associations and anti-associations}
\usage{
extract.fpr.table(X, Y, it.result)
}
\arguments{
\item{X}{a matrix (p x n) where the columns represent the coupled data (i.e., sample)}

\item{Y}{a matrix (q x n) where the columns represent the coupled data (i.e., sample)}

\item{it.result}{a list of the FPR (array of FPRs), wopt (optimal window width), and tau (end scaling factor)}
}
\value{
a list of two dataframes, assoc.fpr.table FPR for each PMCA association and
 antiassoc.fpr.table FPR for each PMCA anti-association.
 row names of each dataframe are  rownames(X) and column names are the rowterms of Y that mapped onto each rowterm of X
}
\description{
Extract dataframes that contain the pairwise FPR associations and anti-associations
}
\examples{
see README
}
