% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.inter.R
\name{get.inter}
\alias{get.inter}
\title{Counts of mappings (and their intersections) for a given list, g, and component, J.}
\usage{
get.inter(g, J = by)
}
\arguments{
\item{g}{a nested list}

\item{J}{component}
}
\value{
an array of dimensions length(g) x length(g)
}
\description{
Counts of mappings (and their intersections) for a given list, g, and component, J.
}
\examples{
see README
}
