% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mapped.fpr.R
\name{get.mapped.fpr}
\alias{get.mapped.fpr}
\title{Produce the lowest pairwise FPR associations and anti-associations}
\usage{
get.mapped.fpr(g, fpr, X)
}
\arguments{
\item{g}{a list that maps rowterms of Y onto rowterms of X for each component}

\item{fpr}{a matrix (p x p-1) of FPRs where the rows represent the rowterms of X and
the columns represent the p-1 components of the covariance matrix related to X}

\item{X}{a matrix (p x n) where the columns represent the coupled data (i.e., sample)}
}
\value{
a dataframe of FPRs for each pairwise relationship contained in g
}
\description{
Produce the lowest pairwise FPR associations and anti-associations
}
\examples{
see README
}
