% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mca.R
\name{get.mca}
\alias{get.mca}
\title{Calculates the matrices needed for PMCA using two coupled datasets}
\usage{
get.mca(X, Y)
}
\arguments{
\item{X}{a matrix (p x n) where the columns represent the coupled data (i.e., sample)}

\item{Y}{a matrix (q x n) where the columns represent the coupled data (i.e., sample)}
}
\value{
a list that contains the components of the singular value decomposition
of the covariance matrix of X and Y (u, sigma, v),
 the principle components (Px) of the covariance matrix related to X, and
 the amplitudes of the co-varying patterns related to X and Y (Zx and Zy),
}
\description{
Calculates the matrices needed for PMCA using two coupled datasets
}
\examples{
see README
}
