% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.scores.R
\name{get.scores}
\alias{get.scores}
\title{Calculates the distance scores}
\usage{
get.scores(Zx, Zy)
}
\arguments{
\item{Zx}{a (p x n+1) matrix the represents the amplitudes of the co-varying patterns in X}

\item{Zy}{a (q x n+1) matrix that represents the amplitudes of the co-varying patterns in Y}
}
\value{
a (q x n x p) array of distance scores of the elements of Zx and Zy.
}
\description{
Calculates the distance scores
}
\examples{
see README
}
