% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative.proc.R
\name{iterative.proc}
\alias{iterative.proc}
\title{Iterative procedure to select the optimal window width}
\usage{
iterative.proc(
  scores,
  alpha = 0.05,
  w,
  tau = 1,
  method = "overall",
  by = 3,
  rowterms,
  plot = TRUE
)
}
\arguments{
\item{scores}{3 dimensional array of the distance scores between Zx and Zy}

\item{alpha}{numeric value for FPR cutoff}

\item{w}{numeric vector of length n of starting window widths for thecomponents}

\item{tau}{numeric value for the scaling value in the window width calculation}

\item{method}{string "overall" or "each" describing the FPR calculation}

\item{by}{numeric integer <= n: that the FPR <= alpha by this component}

\item{rowterms}{vector of string values that denotes names of the elements in Y}

\item{plot}{logical TRUE or FALSE}
}
\value{
a list of the FPR (array of FPRs), wopt (optimal window width), and tau (end scaling factor)
}
\description{
Iterative procedure to select the optimal window width
}
\examples{
see README
}
