% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.graph.R
\name{make.graph}
\alias{make.graph}
\title{Generates a k-partite graph from the mapped results}
\usage{
make.graph(
  dataframe_datalist,
  focus,
  default_threshold = 0.001,
  apply_threshold = NULL,
  debug = FALSE
)
}
\arguments{
\item{focus}{a string specifying the column of interest}

\item{debug}{boolean to show debug statements}

\item{mapped.results}{a list of doubles containing the FPRs for the desired data}

\item{threshold}{a float for the number to threshold on}
}
\value{
a plotly figure
}
\description{
Generates a k-partite graph from the mapped results
}
\examples{
see Visualization_README
}
