% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.patterns.R
\name{match.patterns}
\alias{match.patterns}
\title{Title}
\usage{
match.patterns(score, w, rowterms)
}
\arguments{
\item{score}{an array of distance scores}

\item{w}{a numeric vector denoting the optimal window widths for each component}

\item{rowterms}{a character vector denoting the elements of Y (rownames(Y))}
}
\value{
a list that maps rowterms of Y onto rowterms of X for each component
}
\description{
Title
}
\examples{
see README
}
