% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.proc.R
\name{permutation.proc}
\alias{permutation.proc}
\title{Permutation procedure}
\usage{
permutation.proc(X, Y, method = "overall", B = 1000)
}
\arguments{
\item{X}{a (p x n) matrix}

\item{Y}{a (q by n) matrix}

\item{method}{a character string, "overall" or "each" denoting the FPR method}

\item{B}{an integer denoting the number of permutations}
}
\value{
a 4 dimensional array of distance scores for each of the B permutations
}
\description{
Permutation procedure
}
\examples{
see README
}
