% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.visualization.R
\name{setup.visualization}
\alias{setup.visualization}
\title{Simple wrapper for setup functions}
\usage{
setup.visualization()
}
\description{
Simple wrapper for setup functions
}
\examples{
see README
}
